Change weather service file structure
