+++ services/weather-broadcast/internal/infrastructure/clients/subscription_client.go
