+++ services/weather-broadcast/internal/infrastructure/clients/weather_client.go
