+++ services/weather/internal/domain/models/weather.go
@@ -1,4 +1,4 @@
-package dto
+package models
 
 type (
 	Weather struct {
