+++ services/weather/internal/infrastructure/errors/errors.go
