+++ services/weather/internal/infrastructure/metrics/prometheus.go
