+++ services/weather/internal/infrastructure/providers/openweather_provider.go
@@ -6,37 +6,33 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-	"weather-service/internal/dto"
-	"weather-service/internal/errors"
+	"weather-service/internal/domain/models"
+	"weather-service/internal/infrastructure/errors"
 
 	"weather-forecast/pkg/logger"
 )
 
 type (
-	GetWeatherConditionResponse struct {
-		Text string `json:"text"`
-	}
-
-	GetWeatherCurrentResponse struct {
-		TempC     float64                     `json:"temp_c"`
-		Condition GetWeatherConditionResponse `json:"condition"`
-		Humidity  int                         `json:"humidity"`
-	}
-
-	GetWeatherSuccessResponse struct {
-		Current GetWeatherCurrentResponse `json:"current"`
-	}
-
-	GetWeatherErrorDetails struct {
-		Code    int    `json:"code"`
+	GetOpenWeatherErrorResponse struct {
+		Cod     string `json:"cod"`
 		Message string `json:"message"`
 	}
 
-	GetWeatherErrorResponse struct {
-		Error GetWeatherErrorDetails `json:"error"`
+	GetOpenWeatherMainResponse struct {
+		Temperature float64 `json:"temp"`
+		Humidity    int     `json:"humidity"`
 	}
 
-	WeatherAPIProvider struct {
+	GetOpenWeatherDescriptionResponse struct {
+		Description string `json:"description"`
+	}
+
+	GetOpenWeatherSuccessResponse struct {
+		Weather []GetOpenWeatherDescriptionResponse `json:"weather"`
+		Main    GetOpenWeatherMainResponse          `json:"main"`
+	}
+
+	OpenWeatherProvider struct {
 		apiURL string
 		apiKey string
 		client *http.Client
@@ -44,18 +40,19 @@ type (
 	}
 )
 
-func NewWeatherAPIProvider(apiURL, apiKey string, httpClient *http.Client, logger logger.Logger) *WeatherAPIProvider {
-	return &WeatherAPIProvider{
+func NewOpenWeatherProvider(apiURL, apiKey string, client *http.Client, logger logger.Logger) *OpenWeatherProvider {
+	return &OpenWeatherProvider{
 		apiURL: apiURL,
 		apiKey: apiKey,
-		client: httpClient,
+		client: client,
 		logger: logger,
 	}
 }
 
-func (p *WeatherAPIProvider) GetWeatherByCity(ctx context.Context, city string) (*dto.Weather, error) {
+func (p *OpenWeatherProvider) GetWeatherByCity(ctx context.Context, city string) (*models.Weather, error) {
 
-	const notFoundWeatherAPIErrorCode = 1006
+	const notFoundOpenWeatherErrorCode = "404"
+	const metricUnits = "metric"
 
 	url, err := url.Parse(p.apiURL)
 	if err != nil {
@@ -63,8 +60,9 @@ func (p *WeatherAPIProvider) GetWeatherByCity(ctx context.Context, city string) 
 		return nil, errors.GetWeatherError
 	}
 	queryString := url.Query()
-	queryString.Set("key", p.apiKey)
 	queryString.Set("q", city)
+	queryString.Set("appid", p.apiKey)
+	queryString.Set("units", metricUnits)
 	url.RawQuery = queryString.Encode()
 	stringURL := url.String()
 
@@ -92,36 +90,45 @@ func (p *WeatherAPIProvider) GetWeatherByCity(ctx context.Context, city string) 
 	}
 
 	if resp.StatusCode != http.StatusOK {
-
-		var errResponse GetWeatherErrorResponse
+		var errResponse GetOpenWeatherErrorResponse
 
 		if err := json.Unmarshal(body, &errResponse); err != nil {
 			p.logger.Warnf("Failed to unmarshal response body: %s", err.Error())
 			return nil, errors.GetWeatherError
 		}
 
-		if errResponse.Error.Code == notFoundWeatherAPIErrorCode {
+		if errResponse.Cod == notFoundOpenWeatherErrorCode {
+
 			p.logger.Warnf("City not found: %s", city)
 			return nil, errors.CityNotFoundError
 		} else {
-			p.logger.Warnf("Error from weather api: %s", errResponse.Error.Message)
+			p.logger.Warnf("Error from open weather: %s", errResponse.Message)
 			return nil, errors.GetWeatherError
 		}
 
 	}
 
-	var weather GetWeatherSuccessResponse
+	var weatherResponse GetOpenWeatherSuccessResponse
 
-	if err := json.Unmarshal(body, &weather); err != nil {
+	if err := json.Unmarshal(body, &weatherResponse); err != nil {
 		p.logger.Warnf("Failed to unmarshal response body: %s", err.Error())
 		return nil, errors.GetWeatherError
 	}
 
-	result := dto.Weather{
-		Temperature: weather.Current.TempC,
-		Humidity:    weather.Current.Humidity,
-		Description: weather.Current.Condition.Text,
+	weatherDesc := ""
+
+	if len(weatherResponse.Weather) > 0 {
+		weatherDesc = weatherResponse.Weather[0].Description
+	} else {
+		p.logger.Warnf("OpenWeather did not provide weather description for city: %s", city)
+	}
+
+	result := models.Weather{
+		Temperature: weatherResponse.Main.Temperature,
+		Humidity:    weatherResponse.Main.Humidity,
+		Description: weatherDesc,
 	}
 
 	return &result, nil
+
 }
