+++ services/weather/internal/infrastructure/providers/roundtrip/logging_roundtrip.go
