+++ services/weather/internal/presentation/server/server.go
